Added some more features to the JavaClassGenerator * better recognition of String data types * convert uint8_t to byte * changed unit test * added lcd.h to build

+++ pi4j-core/src/main/java/com/pi4j/wiringpi/WiringShift.java
@@ -45,6 +45,6 @@ public class WiringShift {
     }
 
 
-    public static native int shiftIn (int dPin, int cPin, int order) ;
-    public static native void shiftOut (int dPin, int cPin, int order, int val) ;
+    public static native byte shiftIn (byte dPin, byte cPin, byte order) ;
+    public static native void shiftOut (byte dPin, byte cPin, byte order, byte val) ;
 }
